Stability updates
Reduced initial bboxIP and bboxSL to one variable each.
Updated peak values when performing the semi-auto segmentation (there was an issue where we'd sometimes get a value from the first peak, which would result in no segmentation).
Cleared center of 3d bounding box at top and bottom so we can see the segmentation.